+++ packages/lumx-react/src/components/button/ButtonRoot.tsx
@@ -63,11 +63,6 @@ interface BaseButtonProps extends GenericProps {
      * Whether custom colors are applied to this component.
      */
     useCustomColors?: boolean;
-
-    /**
-     * If the button is rounded or not
-     */
-    isRounded?: boolean;
 }
 
 interface ButtonRootProps extends BaseButtonProps {
@@ -133,7 +128,6 @@ const ButtonRoot: React.FC<ButtonRootProps> = (props) => {
         theme,
         useCustomColors,
         variant,
-        isRounded = false,
         ...forwardedProps
     } = props;
 
@@ -157,7 +151,6 @@ const ButtonRoot: React.FC<ButtonRootProps> = (props) => {
             size,
             theme: emphasis === Emphasis.high && theme,
             variant,
-            isRounded,
         }),
         { [`${CSS_PREFIX}-custom-colors`]: useCustomColors },
     );